requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
yarl==1.18.3

phonenumbers~=9.0.0
email_validator~=2.2.0
pip~=24.3.1
requests~=2.32.3
regex~=2024.11.6
bcrypt~=4.3.0
packaging~=24.2
dnspython~=2.7.0
gunicorn~=23.0.0
PyJWT~=2.10.1
cryptography~=44.0.2
pycparser~=2.22